Make `src/xterm.js` ES2015 and export EventEmitter into its own file
Fix #158
